+++ Logic/Kalender/DsaTool/DSADateCalendarNorbards.cs
@@ -41,8 +41,8 @@ public class DSADateCalendarNorbards : DSADateCalendar {
     public override String getHeadingText()
     {
         int jday = getJDay() - 1;
-        int monthdiv = MathUtil.divisio(jday, DAYS_PER_MONTH);
-        int monthmod = MathUtil.modulo(jday, DAYS_PER_MONTH);
+        int monthdiv = (int)MathUtil.divisio(jday, DAYS_PER_MONTH);
+        int monthmod = (int)MathUtil.modulo(jday, DAYS_PER_MONTH);
 
         string s = String.Format("{0}. Tag im {1}. Mondmonat im {2}. Uh'Jun", monthmod + 1, monthdiv + 1, getYear());
         return s;
